pyproject.toml
[project]
name = "mcp-server-public-transport"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "certifi>=2025.7.14",
    "fastmcp>=2.10.6",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
